requirements.txt
sqlalchemy
uvicorn
fastapi
pydantic
wfdb
smbus2
streamlit
numpy
pandas
scipy
matplotlib
requests
websocket-client
streamlit-lottie
PyJWT
passlib[bcrypt]
bcrypt==3.2.2
twilio
python-dotenv
huggingface_hub
joblib
watchfiles
reportlab
psycopg2-binary>=2.9
